+++ client/app/intake/actions/higherLevelReview.js
@@ -72,17 +72,3 @@ export const submitReview = (intakeId, higherLevelReview) => (dispatch) => {
       }
     );
 };
-
-export const setIssueSelected = (profileDate, issueId, isSelected) => ({
-  type: ACTIONS.SET_ISSUE_SELECTED,
-  payload: {
-    profileDate,
-    issueId,
-    isSelected
-  },
-  meta: {
-    analytics: {
-      label: isSelected ? 'selected' : 'de-selected'
-    }
-  }
-});
