+++ client/app/intake/actions/supplementalClaim.js
@@ -46,17 +46,3 @@ export const submitReview = (intakeId, supplementalClaim) => (dispatch) => {
       }
     );
 };
-
-export const setIssueSelected = (profileDate, issueId, isSelected) => ({
-  type: ACTIONS.SET_ISSUE_SELECTED,
-  payload: {
-    profileDate,
-    issueId,
-    isSelected
-  },
-  meta: {
-    analytics: {
-      label: isSelected ? 'selected' : 'de-selected'
-    }
-  }
-});
