+++ client/app/intake/reducers/higherLevelReview.js
@@ -1,7 +1,7 @@
 import { ACTIONS, REQUEST_STATE, FORM_TYPES } from '../constants';
 import { update } from '../../util/ReducerUtil';
 import { formatDateStr } from '../../util/DateUtil';
-import { getReceiptDateError } from '../util';
+import { getReceiptDateError, formatRatings } from '../util';
 import _ from 'lodash';
 
 const getInformalConferenceError = (responseErrorCodes) => (
@@ -32,6 +32,9 @@ const updateFromServerIntake = (state, serverIntake) => {
     },
     isReviewed: {
       $set: Boolean(serverIntake.receipt_date)
+    },
+    ratings: {
+      $set: state.ratings || formatRatings(serverIntake.ratings)
     },
     isComplete: {
       $set: Boolean(serverIntake.completed_at)
@@ -133,6 +136,20 @@ export const higherLevelReviewReducer = (state = mapDataToInitialHigherLevelRevi
       requestStatus: {
         submitReview: {
           $set: REQUEST_STATE.FAILED
+        }
+      }
+    });
+  case ACTIONS.SET_ISSUE_SELECTED:
+    return update(state, {
+      ratings: {
+        [action.payload.profileDate]: {
+          issues: {
+            [action.payload.issueId]: {
+              isSelected: {
+                $set: action.payload.isSelected
+              }
+            }
+          }
         }
       }
     });
